add compiled application.css for now.  change basePath to id. make call to /publish
