+++ new-form.json
@@ -16,6 +16,9 @@
     }
   ],
   "lists": [
-
+  ],
+  "sections": [],
+  "conditions": [],
+  "fees": [
   ]
 }
